+++ src/pages/MyPage.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { useTheme } from '@emotion/react';
-import { PascalCase } from './Login/PascalCase';
+import { PascalCase } from './Login/UserManagement';
 import { css } from '@emotion/react';
 import { useNavigate } from 'react-router-dom';
